Rest API - Early

+++ controllers/todos.js
@@ -8,15 +8,20 @@ class TodoController
         let data = {title, description, status, due_date};
         Todo.create(data)
         .then(data => res.status(201).json(data))
-        // .catch(err => res.status(400).json(err));
-        .catch(err => res.json(err));
+        .catch(err => 
+        {
+            if(err.name == "SequelizeValidationError")
+                res.status(400).json(err);
+            else
+                res.status(500).json({error : "Internal Server Error"});
+        });
     }
 
     static showAll(req, res)
     {
         Todo.findAll()
         .then(data => res.status(200).json(data))
-        .catch(err => res.status(500).json(err));
+        .catch(err => res.status(500).json({error : "Internal Server Error"}));
     }
 
     static showDatum(req, res)
@@ -24,8 +29,14 @@ class TodoController
         let {id} = req.params;
 
         Todo.findByPk(id)
-        .then(data => res.status(200).json(data))
-        .catch(err => res.status(404).json(err));
+        .then(data => 
+        {
+            if(data)
+                res.status(200).json(data);
+            else
+                res.status(404).json({error : "Data not found"});
+        })
+        .catch(err => res.status(500).json({error : "Internal Server Error"}));
     }
 
     static update(req, res)
@@ -35,8 +46,20 @@ class TodoController
         let {id} = req.params;
         
         Todo.update(data, {where : {id}})
-        .then(() => res.status(200).json(data))
-        .catch(err => res.json(err));
+        .then((value) => 
+        {
+            if(value[0] == [0])
+                res.status(404).json({error : "Data not found"});
+            else
+                res.status(200).json(data);
+        })
+        .catch(err => 
+        {
+            if(err.name == "SequelizeValidationError")
+                res.status(400).json(err);
+            else
+                res.status(500).json({error : "Internal Server Error"});
+        });
     }
 
     static delete(req, res)
@@ -47,11 +70,16 @@ class TodoController
         Todo.findByPk(id)
         .then(value => 
             {
-                data = value;
-                return Todo.destroy({where : {id}})
+                if(value == null)
+                    res.status(404).json({error : "Data not found"});
+                else
+                {
+                    data = value;
+                    return Todo.destroy({where : {id}})
+                    .then(() => res.status(200).json(data))
+                }
             })
-            .then(() => res.status(200).json(data))
-            .catch(err => res.json(err));
+        .catch(err => res.status(500).json({error : "Internal Server Error"}));
     
     }
 }
